Clean up trouble caused by emacs.
Had used emacs to implement the minor edits required for G++ compatibility.
Emacs takes hours and hours of dedicated, trial-and-error attention to
configure its tabbing and alignment options to a sane status. But I want it to
work. Another day. I got another tale of woe from installing Eclipse and trying
to make it "use an existing Makefile" which can be done in more than one way,
but only one way Actually populates the project.

So there is literally 15 minutes of actual editting punctuated by hours of
googling and reinstalling Eclipse. Oracle defaults to installing and
maintaining a 32-bit JRE on a 64-bit Windows OS. That was another cute one.