Bug fixes in Haskell grammar. Layout fixes.
There was a bit of confusion between gconsym and qconsym preventing the
a:b form from being recognized as a list constructor form.

Several Action() functions have been implemented and added to the
grammar. A few, like LPat and LPat3, are just to clarify which
productions are used to match a form. This pointed out a few places
where the prioritized choices had bad priorities.

Also, the layout code now tracks parentheses. If parentheses surround a
domain where curly braces are supposed to be implied, the closing brace
will be inserted before the closing parenthesis. Similar logic could
easily be added for square brackets.

Adding the parenthesis logic to the layout code forestalls the need for
interaction between the parser and the layout required for the 'parse
error' note in the layout specification in the Haskell Report.